+++ KlayGE/Core/Src/Render/RenderLayout.cpp
@@ -63,7 +63,7 @@ namespace KlayGE
 		BOOST_ASSERT(buffer);
 
 		uint32_t size = 0;
-		for (int i = 0; i < vet.size(); ++ i)
+		for (size_t i = 0; i < vet.size(); ++i)
 		{
 			size += vet[i].element_size();
 		}
